Add Valid: 닉네임 , 비밀번호 검증 메소드 클래스로 분리
