+++ src/main/java/study/tdd/simpleboard/api/member/signup/valid/Valid.java
@@ -1,0 +1,5 @@
+package study.tdd.simpleboard.api.member.signup.valid;
+
+public interface Valid {
+    boolean verify(String target);
+}
