+++ src/main/java/study/tdd/simpleboard/api/member/signup/valid/ValidationPassword.java
@@ -1,0 +1,4 @@
+package study.tdd.simpleboard.api.member.signup.valid;
+
+public class ValidationPassword {
+}
